Split link code up into smaller files and adjust re-exports
Smaller, more focused files should make it easier to refactor
AC's link support further. Link based re-exports from Apollo
Client have been adjusted at the same time, to make sure we're
only exporting items that are trully needed by 3rd party links
(and by the `apollo-link` repo).
