+++ src/link/http/__tests__/checkFetcher.ts
@@ -1,0 +1,23 @@
+import { checkFetcher } from '../checkFetcher';
+
+describe('checkFetcher', () => {
+  let oldFetch;
+  beforeEach(() => {
+    oldFetch = window.fetch;
+    delete window.fetch;
+  });
+
+  afterEach(() => {
+    window.fetch = oldFetch;
+  });
+
+  it('throws if no fetch is present', () => {
+    expect(() => checkFetcher(undefined)).toThrow(
+      /fetch is not found globally/,
+    );
+  });
+
+  it('does not throws if no fetch is present but a fetch is passed', () => {
+    expect(() => checkFetcher(() => {})).not.toThrow();
+  });
+});
