+++ src/link/http/__tests__/httpLink.ts
@@ -4,7 +4,8 @@ import Observable from 'zen-observable';
 import { print } from 'graphql';
 
 import { ApolloLink, execute } from '../../core';
-import { HttpLink, createHttpLink } from '../httpLink';
+import { HttpLink } from '../HttpLink';
+import { createHttpLink } from '../createHttpLink';
 
 const sampleQuery = gql`
   query SampleQuery {
@@ -33,29 +34,53 @@ const makeCallback = (done, body) => {
   };
 };
 
-describe('HttpLink', () => {
-  describe('HttpLink-specific tests', () => {
-    it('does not need any constructor arguments', () => {
-      expect(() => new HttpLink()).not.toThrow();
-    });
+const convertBatchedBody = body => {
+  const parsed = JSON.parse(body);
+  return parsed;
+};
 
-    const makePromise = res =>
-      new Promise((resolve, reject) => setTimeout(() => resolve(res)));
+const makePromise =
+  res => new Promise((resolve) => setTimeout(() => resolve(res)));
+
+describe('HttpLink', () => {
+  describe('General', () => {
     const data = { data: { hello: 'world' } };
+    const data2 = { data: { hello: 'everyone' } };
+    const mockError = { throws: new TypeError('mock me') };
+    let subscriber;
 
     beforeEach(() => {
       fetchMock.restore();
-      fetchMock.post('begin:http://data/', makePromise(data));
-      fetchMock.get('begin:http://data/', makePromise(data));
+      fetchMock.post('begin:/data2', makePromise(data2));
+      fetchMock.post('begin:/data', makePromise(data));
+      fetchMock.post('begin:/error', mockError);
+      fetchMock.post('begin:/apollo', makePromise(data));
+
+      fetchMock.get('begin:/data', makePromise(data));
+      fetchMock.get('begin:/data2', makePromise(data2));
+
+      const next = jest.fn();
+      const error = jest.fn();
+      const complete = jest.fn();
+
+      subscriber = {
+        next,
+        error,
+        complete,
+      };
     });
 
     afterEach(() => {
       fetchMock.restore();
     });
 
+    it('does not need any constructor arguments', () => {
+      expect(() => new HttpLink()).not.toThrow();
+    });
+
     it('constructor creates link that can call next and then complete', done => {
       const next = jest.fn();
-      const link = new HttpLink({ uri: 'http://data/' });
+      const link = new HttpLink({ uri: '/data' });
       const observable = execute(link, {
         query: sampleQuery,
       });
@@ -74,7 +99,7 @@ describe('HttpLink', () => {
       const extensions = { myExtension: 'foo' };
 
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
         fetchOptions: { method: 'GET' },
         includeExtensions: true,
       });
@@ -86,29 +111,30 @@ describe('HttpLink', () => {
           expect(body).toBeUndefined();
           expect(method).toBe('GET');
           expect(uri).toBe(
-            'http://data/?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D&extensions=%7B%22myExtension%22%3A%22foo%22%7D',
+            '/data?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D&extensions=%7B%22myExtension%22%3A%22foo%22%7D',
           );
         }),
         error: error => done.fail(error),
       });
     });
 
-    it('supports using a GET request with search and fragment', done => {
+    it('supports using a GET request with search', done => {
       const variables = { params: 'stub' };
 
       const link = createHttpLink({
-        uri: 'http://data/?foo=bar#frag',
+        uri: '/data?foo=bar',
         fetchOptions: { method: 'GET' },
       });
 
       execute(link, { query: sampleQuery, variables }).subscribe({
         next: makeCallback(done, result => {
+          console.log(fetchMock.lastCall());
           const [uri, options] = fetchMock.lastCall();
           const { method, body } = options;
           expect(body).toBeUndefined();
           expect(method).toBe('GET');
           expect(uri).toBe(
-            'http://data/?foo=bar&query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D#frag',
+            '/data?foo=bar&query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D',
           );
         }),
         error: error => done.fail(error),
@@ -118,7 +144,7 @@ describe('HttpLink', () => {
     it('supports using a GET request on the context', done => {
       const variables = { params: 'stub' };
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
       });
 
       execute(link, {
@@ -134,7 +160,7 @@ describe('HttpLink', () => {
           expect(body).toBeUndefined();
           expect(method).toBe('GET');
           expect(uri).toBe(
-            'http://data/?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D',
+            '/data?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D',
           );
         }),
       );
@@ -143,7 +169,7 @@ describe('HttpLink', () => {
     it('uses GET with useGETForQueries', done => {
       const variables = { params: 'stub' };
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
         useGETForQueries: true,
       });
 
@@ -157,7 +183,7 @@ describe('HttpLink', () => {
           expect(body).toBeUndefined();
           expect(method).toBe('GET');
           expect(uri).toBe(
-            'http://data/?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D',
+            '/data?query=query%20SampleQuery%20%7B%0A%20%20stub%20%7B%0A%20%20%20%20id%0A%20%20%7D%0A%7D%0A&operationName=SampleQuery&variables=%7B%22params%22%3A%22stub%22%7D',
           );
         }),
       );
@@ -166,7 +192,7 @@ describe('HttpLink', () => {
     it('uses POST for mutations with useGETForQueries', done => {
       const variables = { params: 'stub' };
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
         useGETForQueries: true,
       });
 
@@ -179,7 +205,7 @@ describe('HttpLink', () => {
           const { method, body } = options;
           expect(body).toBeDefined();
           expect(method).toBe('POST');
-          expect(uri).toBe('http://data/');
+          expect(uri).toBe('/data');
         }),
       );
     });
@@ -187,7 +213,7 @@ describe('HttpLink', () => {
     it('should add client awareness settings to request headers', done => {
       const variables = { params: 'stub' };
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
       });
 
       const clientAwareness = {
@@ -220,7 +246,7 @@ describe('HttpLink', () => {
     it('should not add empty client awareness settings to request headers', done => {
       const variables = { params: 'stub' };
       const link = createHttpLink({
-        uri: 'http://data/',
+        uri: '/data',
       });
 
       const hasOwn = Object.prototype.hasOwnProperty;
@@ -242,101 +268,60 @@ describe('HttpLink', () => {
         }),
       );
     });
-  });
 
-  it("throws for GET if the variables can't be stringified", done => {
-    const link = createHttpLink({
-      uri: 'http://data/',
-      useGETForQueries: true,
-    });
+    it("throws for GET if the variables can't be stringified", done => {
+      const link = createHttpLink({
+        uri: '/data',
+        useGETForQueries: true,
+      });
 
-    let b;
-    const a = { b };
-    b = { a };
-    a.b = b;
-    const variables = {
-      a,
-      b,
-    };
-    execute(link, { query: sampleQuery, variables }).subscribe(
-      result => {
-        done.fail('next should have been thrown from the link');
-      },
-      makeCallback(done, e => {
-        expect(e.message).toMatch(/Variables map is not serializable/);
-        expect(e.parseError.message).toMatch(
-          /Converting circular structure to JSON/,
-        );
-      }),
-    );
-  });
-
-  it("throws for GET if the extensions can't be stringified", done => {
-    const link = createHttpLink({
-      uri: 'http://data/',
-      useGETForQueries: true,
-      includeExtensions: true,
-    });
-
-    let b;
-    const a = { b };
-    b = { a };
-    a.b = b;
-    const extensions = {
-      a,
-      b,
-    };
-    execute(link, { query: sampleQuery, extensions }).subscribe(
-      result => {
-        done.fail('next should have been thrown from the link');
-      },
-      makeCallback(done, e => {
-        expect(e.message).toMatch(/Extensions map is not serializable/);
-        expect(e.parseError.message).toMatch(
-          /Converting circular structure to JSON/,
-        );
-      }),
-    );
-  });
-
-  const convertBatchedBody = body => {
-    const parsed = JSON.parse(body);
-    return parsed;
-  };
-
-  describe('SharedHttpTest', () => {
-    const data = { data: { hello: 'world' } };
-    const data2 = { data: { hello: 'everyone' } };
-    const mockError = { throws: new TypeError('mock me') };
-
-    const makePromise = res =>
-      new Promise((resolve, reject) => setTimeout(() => resolve(res)));
-
-    let subscriber;
-
-    beforeEach(() => {
-      fetchMock.restore();
-      fetchMock.post('begin:/data2', makePromise(data2));
-      fetchMock.post('begin:/data', makePromise(data));
-      fetchMock.post('begin:/error', mockError);
-      fetchMock.post('begin:/apollo', makePromise(data));
-
-      fetchMock.get('begin:/data', makePromise(data));
-      fetchMock.get('begin:/data2', makePromise(data2));
-
-      const next = jest.fn();
-      const error = jest.fn();
-      const complete = jest.fn();
-
-      subscriber = {
-        next,
-        error,
-        complete,
+      let b;
+      const a = { b };
+      b = { a };
+      a.b = b;
+      const variables = {
+        a,
+        b,
       };
+      execute(link, { query: sampleQuery, variables }).subscribe(
+        result => {
+          done.fail('next should have been thrown from the link');
+        },
+        makeCallback(done, e => {
+          expect(e.message).toMatch(/Variables map is not serializable/);
+          expect(e.parseError.message).toMatch(
+            /Converting circular structure to JSON/,
+          );
+        }),
+      );
     });
 
-    afterEach(() => {
-      fetchMock.restore();
+    it("throws for GET if the extensions can't be stringified", done => {
+      const link = createHttpLink({
+        uri: '/data',
+        useGETForQueries: true,
+        includeExtensions: true,
+      });
+
+      let b;
+      const a = { b };
+      b = { a };
+      a.b = b;
+      const extensions = {
+        a,
+        b,
+      };
+      execute(link, { query: sampleQuery, extensions }).subscribe(
+        result => {
+          done.fail('next should have been thrown from the link');
+        },
+        makeCallback(done, e => {
+          expect(e.message).toMatch(/Extensions map is not serializable/);
+          expect(e.parseError.message).toMatch(
+            /Converting circular structure to JSON/,
+          );
+        }),
+      );
     });
 
     it('raises warning if called with concat', () => {
@@ -861,7 +846,7 @@ describe('HttpLink', () => {
     });
   });
 
-  describe('dev warnings', () => {
+  describe('Dev warnings', () => {
     let oldFetch;
     beforeEach(() => {
       oldFetch = window.fetch;
@@ -902,7 +887,7 @@ describe('HttpLink', () => {
     });
   });
 
-  describe('error handling', () => {
+  describe('Error handling', () => {
     let responseBody;
     const text = jest.fn(() => {
       const responseBodyText = '{}';
